Update README.md
A static website for a portfolio is a simple, fast-loading site that showcases an individual's work, skills, and achievements using fixed content. Unlike dynamic websites that require a server-side connection to generate content, static websites are made up of pre-built HTML, CSS, and sometimes JavaScript files that do not change unless manually updated.